requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPdf2
chromadb